requirements.txt
pandas
numpy
seaborn
tqdm
matplotlib
scikit-learn
scipy
dataclasses
catboost
xgboost
lightgbm
dill
ipyleaflet
flask_cors
-e .